address Pani's comments

+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItParameterizedDomain.java
@@ -91,7 +91,6 @@ import static oracle.weblogic.kubernetes.TestConstants.WEBLOGIC_IMAGE_NAME;
 import static oracle.weblogic.kubernetes.TestConstants.WEBLOGIC_IMAGE_TAG;
 import static oracle.weblogic.kubernetes.TestConstants.WEBLOGIC_IMAGE_TO_USE_IN_SPEC;
 import static oracle.weblogic.kubernetes.TestConstants.WEBLOGIC_SLIM;
-import static oracle.weblogic.kubernetes.TestConstants.WEBLOGIC_VERSION;
 import static oracle.weblogic.kubernetes.TestConstants.WLS_DOMAIN_TYPE;
 import static oracle.weblogic.kubernetes.actions.ActionConstants.ARCHIVE_DIR;
 import static oracle.weblogic.kubernetes.actions.ActionConstants.MODEL_DIR;
@@ -122,6 +121,7 @@ import static oracle.weblogic.kubernetes.assertions.TestAssertions.clusterRoleEx
 import static oracle.weblogic.kubernetes.utils.CommonPatchTestUtils.patchDomainResource;
 import static oracle.weblogic.kubernetes.utils.CommonTestUtils.checkPodReady;
 import static oracle.weblogic.kubernetes.utils.CommonTestUtils.checkPodReadyAndServiceExists;
+import static oracle.weblogic.kubernetes.utils.CommonTestUtils.checkServiceDoesNotExist;
 import static oracle.weblogic.kubernetes.utils.CommonTestUtils.createDomainAndVerify;
 import static oracle.weblogic.kubernetes.utils.CommonTestUtils.createImageAndVerify;
 import static oracle.weblogic.kubernetes.utils.CommonTestUtils.createIngressForDomainAndVerify;
@@ -184,6 +184,8 @@ class ItParameterizedDomain {
   private static final String WLDF_OPENSESSION_APP_CONTEXT_ROOT = "opensession";
   private static final String wlSecretName = "weblogic-credentials";
   private static final String DATA_HOME_OVERRIDE = "/u01/oracle/mydata";
+  private static final String miiImageName = "mii-image";
+  private static final String wdtModelFileForMiiDomain = "model-multiclusterdomain-sampleapp-wls.yaml";
 
   private static String opNamespace = null;
   private static String opServiceAccount = null;
@@ -197,8 +199,10 @@ class ItParameterizedDomain {
   private static Domain domainOnPV = null;
   private static int t3ChannelPort = 0;
   private static String miiDomainNamespace = null;
+  private static String miiDomainNegativeNamespace = null;
   private static final String miiDomainUid = "miidomain";
-
+  private static String miiImage = null;
+  private static String encryptionSecretName = "encryptionsecret";
   private String curlCmd = null;
 
   /**
@@ -210,7 +214,7 @@ class ItParameterizedDomain {
    *                   JUnit engine parameter resolution mechanism
    */
   @BeforeAll
-  public static void initAll(@Namespaces(5) List<String> namespaces) {
+  public static void initAll(@Namespaces(6) List<String> namespaces) {
     logger = getLogger();
 
     // get a unique operator namespace
@@ -231,13 +235,18 @@ class ItParameterizedDomain {
     String domainOnPVNamespace = namespaces.get(3);
     assertNotNull(namespaces.get(4));
     String domainInImageNamespace = namespaces.get(4);
+    assertNotNull(namespaces.get(5));
+    miiDomainNegativeNamespace = namespaces.get(5);
 
     // set the service account name for the operator
     opServiceAccount = opNamespace + "-sa";
 
+    // create mii image
+    miiImage = createAndPushMiiImage();
+
     // install and verify operator with REST API
     installAndVerifyOperator(opNamespace, opServiceAccount, true, 0,
-        miiDomainNamespace, domainOnPVNamespace, domainInImageNamespace);
+        miiDomainNamespace, domainOnPVNamespace, domainInImageNamespace, miiDomainNegativeNamespace);
 
     externalRestHttpsPort = getServiceNodePort(opNamespace, "external-weblogic-operator-svc");
 
@@ -247,6 +256,7 @@ class ItParameterizedDomain {
     logger.info("NGINX service name: {0}", nginxServiceName);
     nodeportshttp = getServiceNodePort(nginxNamespace, nginxServiceName, "http");
     logger.info("NGINX http node port: {0}", nodeportshttp);
+
 
     // create model in image domain with multiple clusters
     miiDomain = createMiiDomainWithMultiClusters(miiDomainUid, miiDomainNamespace);
@@ -276,18 +286,26 @@ class ItParameterizedDomain {
       createIngressForDomainAndVerify(domainUid, domainNamespace, nodeportshttp, clusterNameMsPortMap, true,
           true, ADMIN_SERVER_PORT);
     }
+
   }
 
+  /**
+   * Negative test case for creating a model-in-image domain without encryption secret created.
+   * The admin server service/pod will not be created.
+   * Verify the error message should be logged in the operator log.
+   */
   @Test
-  @DisplayName("verify the operator log")
+  @DisplayName("verify the operator log has expected error msg when encryption secret not created for a mii domain")
   public void testOperatorLog() {
+    createMiiDomainNegative("miidomainnegative", miiDomainNegativeNamespace);
     String operatorPodName =
         assertDoesNotThrow(() -> getOperatorPodName(OPERATOR_RELEASE_NAME, opNamespace));
     logger.info("operator pod name: {0}", operatorPodName);
     String operatorPodLog = assertDoesNotThrow(() -> getPodLog(operatorPodName, opNamespace));
     logger.info("operator pod log: {0}", operatorPodLog);
-    assertTrue(operatorPodLog.contains("Introspector Job Log"));
-    assertTrue(operatorPodLog.contains("WebLogic version='" + WEBLOGIC_VERSION + "'"));
+    assertTrue(operatorPodLog.contains(
+        "Domain miidomainnegative is not valid: RuntimeEncryption secret '" + encryptionSecretName
+            + "' not found in namespace '" + miiDomainNegativeNamespace + "'"));
   }
 
   /**
@@ -811,34 +829,37 @@ class ItParameterizedDomain {
    */
   @AfterAll
   public void tearDownAll() {
-
-    // uninstall NGINX release
-    if (nginxHelmParams != null) {
-      assertThat(uninstallNginx(nginxHelmParams))
-          .as("Test uninstallNginx returns true")
-          .withFailMessage("uninstallNginx() did not return true")
-          .isTrue();
-    }
-
-    for (Domain domain : domains) {
-      assertDomainNotNull(domain);
-
-      String domainNamespace = domain.getMetadata().getNamespace();
-
-      // delete cluster role binding created for WLDF policy
-      if (assertDoesNotThrow(
-          () -> clusterRoleBindingExists(domainNamespace + "-" + WLDF_CLUSTER_ROLE_BINDING_NAME))) {
-        assertTrue(deleteClusterRoleBinding(domainNamespace + "-" + WLDF_CLUSTER_ROLE_BINDING_NAME));
+    if (System.getenv("SKIP_CLEANUP") == null
+        || (System.getenv("SKIP_CLEANUP") != null
+        && System.getenv("SKIP_CLEANUP").equalsIgnoreCase("false"))) {
+      // uninstall NGINX release
+      if (nginxHelmParams != null) {
+        assertThat(uninstallNginx(nginxHelmParams))
+            .as("Test uninstallNginx returns true")
+            .withFailMessage("uninstallNginx() did not return true")
+            .isTrue();
       }
-    }
 
-    // delete cluster role created for WLDF policy
-    if (assertDoesNotThrow(() -> clusterRoleExists(WLDF_CLUSTER_ROLE_NAME))) {
-      assertThat(assertDoesNotThrow(() -> deleteClusterRole(WLDF_CLUSTER_ROLE_NAME),
-          "deleteClusterRole failed with ApiException"))
-          .as("Test delete cluster role returns true")
-          .withFailMessage("deleteClusterRole() did not return true")
-          .isTrue();
+      for (Domain domain : domains) {
+        assertDomainNotNull(domain);
+
+        String domainNamespace = domain.getMetadata().getNamespace();
+
+        // delete cluster role binding created for WLDF policy
+        if (assertDoesNotThrow(
+            () -> clusterRoleBindingExists(domainNamespace + "-" + WLDF_CLUSTER_ROLE_BINDING_NAME))) {
+          assertTrue(deleteClusterRoleBinding(domainNamespace + "-" + WLDF_CLUSTER_ROLE_BINDING_NAME));
+        }
+      }
+
+      // delete cluster role created for WLDF policy
+      if (assertDoesNotThrow(() -> clusterRoleExists(WLDF_CLUSTER_ROLE_NAME))) {
+        assertThat(assertDoesNotThrow(() -> deleteClusterRole(WLDF_CLUSTER_ROLE_NAME),
+            "deleteClusterRole failed with ApiException"))
+            .as("Test delete cluster role returns true")
+            .withFailMessage("deleteClusterRole() did not return true")
+            .isTrue();
+      }
     }
   }
 
@@ -850,23 +871,8 @@ class ItParameterizedDomain {
    */
   private static Domain createMiiDomainWithMultiClusters(String domainUid, String domainNamespace) {
 
-    final String miiImageName = "mii-image";
-    final String wdtModelFileForMiiDomain = "model-multiclusterdomain-sampleapp-wls.yaml";
-
     // admin/managed server name here should match with WDT model yaml file
     String adminServerPodName = domainUid + "-" + ADMIN_SERVER_NAME_BASE;
-
-    // create image with model files
-    logger.info("Creating image with model file {0} and verify", wdtModelFileForMiiDomain);
-    List<String> appSrcDirList = new ArrayList<>();
-    appSrcDirList.add(MII_BASIC_APP_NAME);
-    appSrcDirList.add(WLDF_OPENSESSION_APP);
-    String miiImage =
-        createMiiImageAndVerify(miiImageName, Collections.singletonList(MODEL_DIR + "/" + wdtModelFileForMiiDomain),
-            appSrcDirList, WEBLOGIC_IMAGE_NAME, WEBLOGIC_IMAGE_TAG, WLS_DOMAIN_TYPE, false);
-
-    // docker login and push image to docker registry if necessary
-    dockerLoginAndPushImageToRegistry(miiImage);
 
     // create docker registry secret to pull the image from registry
     // this secret is used only for non-kind cluster
@@ -880,7 +886,6 @@ class ItParameterizedDomain {
 
     // create encryption secret
     logger.info("Creating encryption secret");
-    String encryptionSecretName = "encryptionsecret";
     createSecretWithUsernamePassword(encryptionSecretName, domainNamespace, "weblogicenc", "weblogicenc");
 
     // construct the cluster list used for domain custom resource
@@ -1539,5 +1544,104 @@ class ItParameterizedDomain {
                 condition.getRemainingTimeInMS()))
         .until(assertDoesNotThrow(() -> fileExistsInPod(namespace, podName, fileName),
             "fileExistsInPod failed with IOException, ApiException or InterruptedException"));
+  }
+
+  /**
+   * Negative test case for creating a model-in-image domain without encryption secret created.
+   * The admin server service/pod will not be created.
+   * The error message should be logged in the operator log.
+   *
+   * @param domainNamespace namespace in which the domain will be created
+   * @return oracle.weblogic.domain.Domain objects
+   */
+  private Domain createMiiDomainNegative(String domainUid, String domainNamespace) {
+
+    // admin/managed server name here should match with WDT model yaml file
+    String adminServerPodName = domainUid + "-" + ADMIN_SERVER_NAME_BASE;
+
+    // create docker registry secret to pull the image from registry
+    // this secret is used only for non-kind cluster
+    logger.info("Creating docker registry secret in namespace {0}", domainNamespace);
+    createOcirRepoSecret(domainNamespace);
+
+    // create secret for admin credentials
+    logger.info("Creating secret for admin credentials");
+    String adminSecretName = "weblogic-credentials";
+    createSecretWithUsernamePassword(adminSecretName, domainNamespace, ADMIN_USERNAME_DEFAULT, ADMIN_PASSWORD_DEFAULT);
+
+    // create the domain CR without encryption secret created
+    Domain domain = new Domain()
+        .apiVersion(DOMAIN_API_VERSION)
+        .kind("Domain")
+        .metadata(new V1ObjectMeta()
+            .name(domainUid)
+            .namespace(domainNamespace))
+        .spec(new DomainSpec()
+            .domainUid(domainUid)
+            .domainHome("/u01/domains/" + domainUid)
+            .domainHomeSourceType("FromModel")
+            .image(miiImage)
+            .addImagePullSecretsItem(new V1LocalObjectReference()
+                .name(OCIR_SECRET_NAME))
+            .webLogicCredentialsSecret(new V1SecretReference()
+                .name(adminSecretName)
+                .namespace(domainNamespace))
+            .includeServerOutInPodLog(true)
+            .serverStartPolicy("IF_NEEDED")
+            .serverPod(new ServerPod()
+                .addEnvItem(new V1EnvVar()
+                    .name("JAVA_OPTIONS")
+                    .value("-Dweblogic.StdoutDebugEnabled=false"))
+                .addEnvItem(new V1EnvVar()
+                    .name("USER_MEM_ARGS")
+                    .value("-Djava.security.egd=file:/dev/./urandom ")))
+            .adminServer(new AdminServer()
+                .serverStartState("RUNNING")
+                .adminService(new AdminService()
+                    .addChannelsItem(new Channel()
+                        .channelName("default")
+                        .nodePort(0))))
+            .configuration(new Configuration()
+                .model(new Model()
+                    .domainType(WLS_DOMAIN_TYPE)
+                    .runtimeEncryptionSecret(encryptionSecretName))));
+
+    setPodAntiAffinity(domain);
+
+    // create model in image domain
+    logger.info("Creating model in image domain {0} in namespace {1} using docker image {2}",
+        domainUid, domainNamespace, miiImage);
+    createDomainAndVerify(domain, domainNamespace);
+
+    // sleep for a while and check that the admin server service does not exist since the encryption secret not created
+    try {
+      Thread.sleep(60000);
+    } catch (InterruptedException ie) {
+      // ignore
+    }
+    checkServiceDoesNotExist(adminServerPodName, domainNamespace);
+
+    return domain;
+  }
+
+  /**
+   * Create mii image and push it to the registry.
+   *
+   * @return mii image created
+   */
+  private static String createAndPushMiiImage() {
+    // create image with model files
+    logger.info("Creating image with model file {0} and verify", wdtModelFileForMiiDomain);
+    List<String> appSrcDirList = new ArrayList<>();
+    appSrcDirList.add(MII_BASIC_APP_NAME);
+    appSrcDirList.add(WLDF_OPENSESSION_APP);
+    miiImage =
+        createMiiImageAndVerify(miiImageName, Collections.singletonList(MODEL_DIR + "/" + wdtModelFileForMiiDomain),
+            appSrcDirList, WEBLOGIC_IMAGE_NAME, WEBLOGIC_IMAGE_TAG, WLS_DOMAIN_TYPE, false);
+
+    // docker login and push image to docker registry if necessary
+    dockerLoginAndPushImageToRegistry(miiImage);
+
+    return miiImage;
   }
 }
